Napravljena soft delete operacija za UserEntity.

+++ be_t4_04_project/src/main/java/com/ikt/t4/project/repositories/UserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ikt.t4.project.entities.UserEntity;
 
-public interface UserRepository extends CrudRepository<UserEntity, Integer> {
+public interface UserRepository extends CrudRepository<UserEntity, Long> {
 	Boolean existsByUsername(String username);
 }
